pyproject.toml
[project]
name = "suno-musical-box"
version = "0.1.0"
description = "Biometric-driven music generation using Colmi R02 rings and Suno API"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary,pool]>=3.1",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pygame>=2.5",
    "python-dotenv>=1.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
ble = [
    "bleak>=0.21",
]
